--- src/main/java/Neat/and/Tidy/Solutions/cleaning/service/app/service/AdminServiceImpl.java
@@ -1,9 +1,0 @@
-package Neat.and.Tidy.Solutions.cleaning.service.app.service;
-
-import Neat.and.Tidy.Solutions.cleaning.service.app.data.dto.request.*;
-import Neat.and.Tidy.Solutions.cleaning.service.app.data.dto.response.*;
-
-public class AdminServiceImpl implements AdminService{
-
-
-}
