--- src/main/java/Neat/and/Tidy/Solutions/cleaning/service/app/service/OrderServiceImpl.java
@@ -1,4 +1,0 @@
-package Neat.and.Tidy.Solutions.cleaning.service.app.service;
-
-public class OrderServiceImpl {
-}
