+++ src/main/java/Neat/and/Tidy/Solutions/cleaning/service/app/service/impl/AdminServiceImpl.java
@@ -1,0 +1,10 @@
+package Neat.and.Tidy.Solutions.cleaning.service.app.service.impl;
+
+import Neat.and.Tidy.Solutions.cleaning.service.app.data.dto.request.*;
+import Neat.and.Tidy.Solutions.cleaning.service.app.data.dto.response.*;
+import Neat.and.Tidy.Solutions.cleaning.service.app.service.AdminService;
+
+public class AdminServiceImpl implements AdminService {
+
+
+}
