+++ src/main/java/Neat/and/Tidy/Solutions/cleaning/service/app/service/impl/OrderServiceImpl.java
@@ -1,0 +1,4 @@
+package Neat.and.Tidy.Solutions.cleaning.service.app.service.impl;
+
+public class OrderServiceImpl {
+}
